% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddIntervalsToFolder.R
\name{AddIntervalsToFolder}
\alias{AddIntervalsToFolder}
\title{AddIntervalsToFolder}
\usage{
AddIntervalsToFolder(ListCorrelation, intervall = 0.05)
}
\arguments{
\item{ListCorrelation}{folder structure with a depth on 1.}

\item{intervall}{percentage distance to the apex in which all data are to be included.}
}
\value{
Returns the folder structure with intervals addet.
}
\description{
Adds intervalls to a folder structure with a depth on 1.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kröger
}
