% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelLoader.R
\name{AWIExcelLoader}
\alias{AWIExcelLoader}
\title{AWIExcelLoader}
\usage{
AWIExcelLoader(
  Excelname,
  AgeTxtName = NULL,
  Tables = c("Organic", "Element", "GrainSize", "Mineral", "Diatom"),
  FixFormat = F
)
}
\arguments{
\item{Excelname}{Excel sheet from the AWI database.}

\item{AgeTxtName}{Age textfile from Bacon with a d.by from 0.25. This is optional.}

\item{Tables}{Selecting which table sheets to load.}

\item{FixFormat}{standardizes all data so that it is easier to calculate with them in the future.}
}
\value{
AWIExcelLoader retruns a List of all interpolatet excel sheets.
}
\description{
Loads an Excel of the AWI database and interpolates it.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kröger
}
