% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeForBoris.R
\name{CodeforBoris}
\alias{CodeforBoris}
\title{CodeforBoris}
\usage{
CodeforBoris()
}
\value{
Return perfect code.
}
\description{
Code Sampels for Boris.
}
\note{
This function has only been developed for Boris Biskaborn.
}
\author{
Tim Kröger
}
