% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterDiscard.R
\name{FilterDiscard}
\alias{FilterDiscard}
\title{FilterDiscard}
\usage{
FilterDiscard(data, filter = c(0, 0))
}
\arguments{
\item{data}{Dataframe from LoadMultivar.}

\item{filter}{Vector of 2 Integer. First Number stands for min Precent. Second Number stands for minimum number of rows with "min percent".}
}
\value{
Returns a Dataframe.
}
\description{
Checks how much data would be lost due to filtering.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kröger
}
