% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folderCreater.R
\name{FolderCreater}
\alias{FolderCreater}
\title{FolderCreater}
\usage{
FolderCreater(data, format = "collective", SegmentNames = "other")
}
\arguments{
\item{data}{Data sheets in with ages and values to be sorted in.}

\item{format}{Information about the format of the data sheets:\cr\cr
If format="collective" it is expected that the first column contains an age and all other columns contain the corresponding values.\cr\cr
If format="segment" it is expected that there are always alternating ages and values next to each other.\cr\cr
Starting with the age.\cr
Example: Age...V1...Age...V2...}

\item{SegmentNames}{The name of the segments if format="segment" under which they will be listed.}
}
\value{
Returns a folder structure based on lists.
}
\description{
Assembles lists of data tables into a manageable folder sythem.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kröger
}
