% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpolateVector.R
\name{InterpolateVector}
\alias{InterpolateVector}
\title{InterpolateVector}
\usage{
InterpolateVector(age, data, thick = 50)
}
\arguments{
\item{age}{Vector of ages from the date.}

\item{data}{Vector of one parameter.}

\item{thick}{thickness of ages to interpolate to.}
}
\value{
Return an interpolatet vector of the given date.
}
\description{
Interplolates a vector
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kröger
}
