% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListCorrelation.R
\name{ListCorrelation}
\alias{ListCorrelation}
\title{ListCorrelation}
\usage{
ListCorrelation(Main, List, thick = 50, offset = 2000)
}
\arguments{
\item{Main}{Vector that consists of an age and data vector. c(Age,Vector).}

\item{List}{List of Value from FolderCreater.}

\item{thick}{Thickness of age intervals to interpolate to.}

\item{offset}{Size of the area in which all correlations are calculated. must be dividable by the thickness!}
}
\value{
Returns a list of all correlations in the offset area.
}
\description{
Calculates with two vectors all correlations with time lag.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kröger
}
