% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadMultivar.R
\name{LoadMultivar}
\alias{LoadMultivar}
\title{LoadMultivar}
\usage{
LoadMultivar(
  data,
  sep = ",",
  DataAsPercent = T,
  SartOfData = 4,
  Comparison = 2
)
}
\arguments{
\item{data}{Name of the dataset in CSV format.}

\item{sep}{Separator in the CSV.}

\item{DataAsPercent}{Bool indicating whether data should be convertet to percent or not.}

\item{SartOfData}{Number indicating the Collum where the Diatom data stats.}

\item{Comparison}{Number indicating the Collum where the Depth or Age is located.}
}
\value{
Returns a Dataframe.
}
\description{
Loads Diatom data for Multivariate statistics
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kröger
}
