% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoessFromData.R
\name{LoessFromData}
\alias{LoessFromData}
\title{LoessFromData}
\usage{
LoessFromData(data, alpha = 0.05, span = 0.2, type = "vector")
}
\arguments{
\item{data}{data from the Folder structure or the whole Folder structur at once.}

\item{alpha}{size of the confidence intervals.}

\item{span}{the parameter α which controls the degree of smoothing.}

\item{type}{decides whether to read the whole fodler or only a sublist.\cr type="vector" for a list \cr type=="folder" for the whole folder}
}
\value{
AWIExcelLoader retruns a List of all interpolatet excel sheets.
}
\description{
Creates a List from the new calibrated vectors of the Loess function with confidence intervals.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kröger
}
