% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoessPlotFolder.R
\name{LoessPlotFolder}
\alias{LoessPlotFolder}
\title{LoessPlotFolder}
\usage{
LoessPlotFolder(
  data,
  alpha = 0.05,
  span = 0.2,
  name = "Set_main",
  custom_age = NA
)
}
\arguments{
\item{data}{List of named data or a List List of named data.}

\item{alpha}{size of the confidence intervals.}

\item{span}{the parameter α which controls the degree of smoothing.}

\item{name}{Name for the Plot when one the direct data was used.}

\item{custom_age}{custom y Achses max age.}
}
\value{
Plots data.
}
\description{
Plots data from Folder Structure with the Loess function.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kröger
}
