% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDeviation.R
\name{PlotDeviation}
\alias{PlotDeviation}
\title{PlotDeviation}
\usage{
PlotDeviation(data, thick, offset, plotName = "", colors = c("black"))
}
\arguments{
\item{data}{Combind folder structur from the ListCorrelation and AddIntervalsToFolder function.}

\item{thick}{Thickness of age intervals from the functions bervore.}

\item{offset}{Size of the area in which all correlations are calculated from the functions bervore.}

\item{plotName}{Name of the Plot.}

\item{colors}{Colors of the Groups. One COlor for each group needed.}
}
\value{
Retruns PlotDeviation. Legend is displayed as a text file in the console
}
\description{
Plots deviation from the combined results from the ListCorrelation and AddIntervalsToFolder function.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kröger
}
