% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFolder.R
\name{PlotFolder}
\alias{PlotFolder}
\title{PlotFolder}
\usage{
PlotFolder(data, name = "Set_main")
}
\arguments{
\item{data}{List of named data or a List List of named data.}

\item{name}{Name for the Plot when one the direct data was used..}
}
\value{
Plots data.
}
\description{
Plots data from Folder Structure.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kröger
}
