% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resample.R
\name{Resample}
\alias{Resample}
\title{Resample}
\usage{
Resample(SourceData, Datalist)
}
\arguments{
\item{SourceData}{Dataframe after the other dataframes are to be resampled.}

\item{Datalist}{A list of dataframes to be resampled.}
}
\value{
Resample retruns a List of all all resampled dataframes as a List.
}
\description{
Resample a List of Data to an origin data set.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kröger
}
