% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testfunction.R
\name{TestIntToString}
\alias{TestIntToString}
\title{Test}
\usage{
TestIntToString(int)
}
\arguments{
\item{int}{Integer}
}
\value{
Return a String of the Int
}
\description{
Function for Testing Packages
}
\note{
Only for Testing
}
\author{
Tim Kröger
}
