% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VectorCorrelation.R
\name{VectorCorrelation}
\alias{VectorCorrelation}
\title{VectorCorrelation}
\usage{
VectorCorrelation(x, y, offset = 0)
}
\arguments{
\item{x}{List from the function VectorCorrelation.}

\item{y}{List from the function VectorCorrelation.}

\item{offset}{Offset of ages. Must be in limits of the interpolation.}
}
\value{
Retruns the r value from the correlation.
}
\description{
Correlation of two vectors with optional offset.
}
\note{
This function has only been developed for the Alfred Wegener Institute Helmholtz Centre for Polar and Marine Research and should therefore only be used in combination with their database.
}
\author{
Tim Kröger
}
